start web and test

+++ src/workers/inc/base_workers_repository.h
@@ -5,7 +5,7 @@
 
 class BaseWorkersRepository {
  public:
-  virtual int create(const WorkerPost& worker) = 0;
+  virtual size_t create(const WorkerPost& worker) = 0;
 
   virtual WorkerBaseInf read(size_t worker_id) = 0;
   virtual size_t workers_count() = 0;
